refactor: organize interfaces and types for better maintainability
- Moved shared interfaces to main types/index.ts
- Created dedicated types/contact.ts for contact-specific types
- Unified ContactForm and ContactFormData interfaces
- Updated all imports to use shared types from central location
- Enhanced type reusability and maintainability
- All type definitions now follow consistent patterns
