+++ src/hooks/index.ts
@@ -1,11 +1,13 @@
-export {
-  useCSRFSecurity,
-  type SecurityError,
-  type UseCSRFSecurityReturn,
-} from "./useCSRFSecurity";
+export { useCSRFSecurity, type UseCSRFSecurityReturn } from "./useCSRFSecurity";
 
 export {
   useContactSubmission,
-  type ContactFormData,
   type UseContactSubmissionReturn,
 } from "./useContactSubmission";
+
+// Re-export commonly used types from main types file
+export type {
+  ContactFormData,
+  SecurityError,
+  CSRFTokenResponse,
+} from "@/types";
